+++ lib/openstudio/lib/ami_list.rb
@@ -21,6 +21,8 @@
 # Class for managing the AMI ids based on the openstudio version and the openstudio-server version
 
 class OpenStudioAmis
+  include Logging
+
   VALID_OPTIONS = [
     :openstudio_version, :openstudio_server_version, :host, :url, :stable
   ]
@@ -100,6 +102,7 @@ class OpenStudioAmis
     json[version]
   end
 
+  # Return the AMIs for the server and worker. Version 2 also does a lookup of the stable version
   def get_ami_version_2
     json = list
 
@@ -113,10 +116,34 @@ class OpenStudioAmis
       amis = value[:amis]
     elsif @options[:openstudio_version] != 'default'
       if @options[:stable]
-        stable = json[:openstudio][@options[:openstudio_version].to_sym][:stable]
-        fail "Could not find a stable version for openstudio version #{@options[:openstudio_version]}" unless stable
-        value = json[:openstudio][@options[:openstudio_version].to_sym][stable.to_sym]
-        amis = value[:amis]
+        stable = nil
+        if json[:openstudio][@options[:openstudio_version].to_sym]
+          stable = json[:openstudio][@options[:openstudio_version].to_sym][:stable]
+        end
+
+        if stable
+          value = json[:openstudio][@options[:openstudio_version].to_sym][stable.to_sym]
+          amis = value[:amis]
+        else
+          logger.info "Could not find a stable version for OpenStudio version #{@options[:openstudio_version]}. "\
+                      'Looking up older versions to find the latest stable.' unless stable
+
+          json[:openstudio].each do |os_version, values|
+            next if os_version == :default
+            if values.key? :stable
+              # don't check versions newer than what we are requesting
+              next if os_version.to_s.to_version > @options[:openstudio_version].to_s.to_version
+              stable = json[:openstudio][os_version][:stable]
+              logger.info "Found a stable version for OpenStudio version #{os_version} with OpenStudio Server version #{stable}"
+              value = values[stable.to_sym]
+              amis = value[:amis]
+
+              break
+            end
+          end
+
+          fail "Could not find a stable version for openstudio version #{@options[:openstudio_version]}" unless amis
+        end
       else
         # return the default version (which is the latest)
         default = json[:openstudio][@options[:openstudio_version].to_sym][:default]
@@ -125,6 +152,8 @@ class OpenStudioAmis
         amis = value[:amis]
       end
     end
+
+    logger.info "AMI IDs are #{amis}" if amis
 
     amis
   end
@@ -138,6 +167,7 @@ class OpenStudioAmis
 
     url = URI.parse(uri_str)
     req = Net::HTTP::Get.new(url.path)
+    logger.info "Fetching AMI list from #{uri_str}"
     response = Net::HTTP.start(url.host, url.port) { |http| http.request(req) }
     case response
       when Net::HTTPSuccess then