+++ lib/openstudio/lib/openstudio_aws_wrapper.rb
@@ -1,6 +1,5 @@
-# NOTE: Do not modify this file as it is copied over. Modify the source file and rerun rake import_files
 ######################################################################
-#  Copyright (c) 2008-2014, Alliance for Sustainable Energy.
+#  Copyright (c) 2008-2015, Alliance for Sustainable Energy.
 #  All rights reserved.
 #
 #  This library is free software; you can redistribute it and/or
@@ -55,12 +54,22 @@ class OpenStudioAwsWrapper
 
   def initialize(options = {}, group_uuid = nil)
     @group_uuid = group_uuid || (SecureRandom.uuid).gsub('-', '')
+    logger.info "GroupUUID is #{@group_uuid}"
 
     @security_groups = []
     @key_pair_name = nil
     @private_key_file_name = nil
+    @region = options[:region] || 'unknown-region'
 
-    @worker_keys = SSHKey.generate
+    # If the keys exist in the directory then load those, otherwise create new ones.
+    work_dir = options[:save_directory] || '.'
+    if File.exist?(File.join(work_dir, 'ec2_worker_key.pem')) && File.exist?(File.join(work_dir, 'ec2_worker_key.pub'))
+      logger.info "Worker keys already exist, loading from #{work_dir}"
+      load_worker_key(File.join(work_dir, 'ec2_worker_key.pem'))
+    else
+      logger.info 'Generating new worker keys'
+      @worker_keys = SSHKey.generate
+    end
 
     @private_key = nil # Private key data
 
@@ -118,15 +127,50 @@ class OpenStudioAwsWrapper
     describe_availability_zones.to_json
   end
 
-  def describe_total_instances
+  def total_instances_count
     resp = @aws.describe_instance_status
 
-    region = resp.instance_statuses.length > 0 ? resp.instance_statuses.first.availability_zone : 'no_instances'
-    { total_instances: resp.instance_statuses.length, region: region }
+    availability_zone = resp.instance_statuses.length > 0 ? resp.instance_statuses.first.availability_zone : 'no_instances'
+
+    { total_instances: resp.instance_statuses.length, region: @region, availability_zone: availability_zone }
   end
 
-  def describe_total_instances_json
-    describe_total_instances.to_json
+  def describe_all_instances
+    resp = @aws.describe_instance_status
+
+    resp
+  end
+
+  # describe the instances by group id (this is the default method)
+  def describe_instances
+    resp = nil
+    if group_uuid
+      resp = @aws.describe_instances(
+        filters: [
+          # {name: 'instance-state-code', values: [0.to_s, 16.to_s]}, # running or pending -- any state
+          { name: 'tag-key', values: ['GroupUUID'] },
+          { name: 'tag-value', values: [group_uuid.to_s] }
+        ]
+      )
+    else
+      resp = @aws.describe_instances
+    end
+
+    # Any additional filters
+    instance_data = nil
+    if resp
+      instance_data = []
+      resp.reservations.each do |r|
+        r.instances.each do |i|
+          i_h = i.to_hash
+          if i_h[:tags].any? { |h| (h[:key] == 'GroupUUID') && (h[:value] == group_uuid.to_s) }
+            instance_data << i_h
+          end
+        end
+      end
+    end
+
+    instance_data
   end
 
   # return all of the running instances, or filter by the group_uuid & instance type
@@ -220,13 +264,14 @@ class OpenStudioAwsWrapper
   end
 
   def terminate_instances(ids)
+    resp = nil
     begin
       resp = @aws.terminate_instances(
         instance_ids: ids
       )
     rescue Aws::EC2::Errors::InvalidInstanceIDNotFound
       # Log that the instances couldn't be found?
-      return resp = { error: 'instances could not be found' }
+      resp = { error: 'instances could not be found' }
     end
 
     resp
@@ -263,12 +308,26 @@ class OpenStudioAwsWrapper
     logger.info("create key pair: #{@key_pair_name}")
   end
 
+  # Delete the key pair from aws
+  def delete_key_pair(key_pair_name = nil)
+    tmp_name = key_pair_name || "os-key-pair-#{@group_uuid}"
+    resp = nil
+    begin
+      logger.info "Trying to delete key pair #{tmp_name}"
+      resp = @aws.delete_key_pair(key_name: tmp_name)
+    rescue
+      logger.info "could not delete the key pair '#{tmp_name}'"
+    end
+
+    resp
+  end
+
   def load_private_key(filename)
     unless File.exist? filename
       # check if the file basename exists in your user directory
       filename = File.expand_path("~/.ssh/#{File.basename(filename)}")
       if File.exist? filename
-        puts "Found key of same name in user's home ssh folder #{filename}"
+        logger.info "Found key of same name in user's home ssh folder #{filename}"
         # using the key in your home directory
       else
         fail "Could not find private key #{filename}" unless File.exist? filename
@@ -279,11 +338,24 @@ class OpenStudioAwsWrapper
     @private_key = File.read(filename)
   end
 
-  def save_private_key(filename)
+  # Load the worker key for communicating between the server and worker instances on AWS. The public key
+  # will be automatically created when loading the private key
+  #
+  # @param private_key_filename [String] Fully qualified path to the worker private key
+  def load_worker_key(private_key_filename)
+    logger.info "Loading worker keys from #{private_key_filename}"
+    @worker_keys_filename = private_key_filename
+    @worker_keys = SSHKey.new(File.read(@worker_keys_filename))
+  end
+
+  # Save the private key to disk
+  def save_private_key(directory = '.', filename = 'ec2_server_key.pem')
     if @private_key
-      @private_key_file_name = File.expand_path filename
-      File.open(filename, 'w') { |f| f << @private_key }
-      File.chmod(0600, filename)
+      @private_key_file_name = File.expand_path "#{directory}/#{filename}"
+      logger.info "Saving server private key in #{@private_key_file_name}"
+      File.open(@private_key_file_name, 'w') { |f| f << @private_key }
+      logger.info 'Setting permissions of server private key to 0600'
+      File.chmod(0600, @private_key_file_name)
     else
       fail "No private key found in which to persist with filename #{filename}"
     end
@@ -291,8 +363,15 @@ class OpenStudioAwsWrapper
 
   # save off the worker public/private keys that were created
   def save_worker_keys(directory = '.')
-    File.open("#{directory}/ec2_worker_key.pem", 'w') { |f| f << @worker_keys.private_key }
-    File.open("#{directory}/ec2_worker_key.pub", 'w') { |f| f << @worker_keys.public_key }
+    @worker_keys_filename = "#{directory}/ec2_worker_key.pem"
+    logger.info "Saving worker private key in #{@worker_keys_filename}"
+    File.open(@worker_keys_filename, 'w') { |f| f << @worker_keys.private_key }
+    logger.info 'Setting permissions of worker private key to 0600'
+    File.chmod(0600, @worker_keys_filename)
+
+    wk = "#{directory}/ec2_worker_key.pub"
+    logger.info "Saving worker public key in #{wk}"
+    File.open(wk, 'w') { |f| f << @worker_keys.public_key }
   end
 
   def launch_server(image_id, instance_type, launch_options = {})
@@ -324,8 +403,7 @@ class OpenStudioAwsWrapper
       user_id: 'unknown_user',
       tags: [],
       ebs_volume_size: nil,
-      availability_zone: @server.data.availability_zone,
-      worker_public_key: ''
+      availability_zone: @server.data.availability_zone
     }
     launch_options = defaults.merge(launch_options)
 
@@ -336,7 +414,6 @@ class OpenStudioAwsWrapper
     logger.info("worker user_data #{user_data.inspect}")
 
     # thread the launching of the workers
-
     num.times do
       @workers << OpenStudioAwsInstance.new(@aws, :worker, @key_pair_name, @security_groups, @group_uuid,
                                             @private_key, @private_key_file_name, @proxy)
@@ -391,14 +468,19 @@ class OpenStudioAwsWrapper
   end
 
   # method to query the amazon api to find the server (if it exists), based on the group id
-  # if it is found, then it will set the @server member variable.
+  # if it is found, then it will set the @server instance variable. The security groups are assigned from the
+  # server node information on AWS if the security groups have not been initialized yet.
+  #
   # Note that the information around keys and security groups is pulled from the instance information.
+  # @param server_data_hash [Hash] Server data
+  # @option server_data_hash [String] :group_id Group ID of the analysis
+  # @option server_data_hash [String] :server.private_key_file_name Name of the private key to communicate to the server
   def find_server(server_data_hash)
     @group_uuid = server_data_hash[:group_id] || @group_uuid
     load_private_key(server_data_hash[:server][:private_key_file_name])
 
-    logger.info "finding the server for groupid of #{group_uuid}"
-    fail 'no group uuid defined either in member variable or method argument' if group_uuid.nil?
+    logger.info "Finding the server for GroupUUID of #{group_uuid}"
+    fail 'no GroupUUID defined either in member variable or method argument' if @group_uuid.nil?
 
     # This should really just be a single call to describe running instances
     @server = nil
@@ -409,7 +491,17 @@ class OpenStudioAwsWrapper
       if !@server
         if resp
           logger.info "Server found and loading data into object [instance id is #{resp[:instance_id]}]"
-          @server = OpenStudioAwsInstance.new(@aws, :server, resp[:key_name], resp[:security_groups].first[:group_name], group_uuid, @private_key, @private_key_file_name, @proxy)
+
+          sg = resp[:security_groups].map { |s| s[:group_id] }
+          # Set the security groups of the object if these groups haven't been assigned yet.
+          @security_groups = sg if @security_groups.empty?
+          logger.info "The security groups in aws wrapper are #{@security_groups}"
+
+          # set the key name from AWS if it isn't yet assigned
+          logger.info 'Setting the keyname in the aws wrapper'
+          @key_pair_name = resp[:key_name] unless @key_pair_name
+
+          @server = OpenStudioAwsInstance.new(@aws, :server, @key_pair_name, sg, @group_uuid, @private_key, @private_key_file_name, @proxy)
 
           @server.load_instance_data(resp)
         end
@@ -417,21 +509,25 @@ class OpenStudioAwsWrapper
         logger.info "Server instance is already defined with instance #{resp[:instance_id]}"
       end
     else
-      puts 'could not find a running server instance'
+      logger.info 'could not find a running server instance'
     end
 
-    # find the workers
+    # Find the worker instances.
     if @workers.size == 0
       resp = describe_running_instances(group_uuid, :worker)
       if resp
         resp.each do |r|
-          @workers << OpenStudioAwsInstance.new(@aws, :worker, r[:key_name], r[:security_groups].first[:group_name], group_uuid, @private_key, @private_key_file_name, @proxy)
+          @workers << OpenStudioAwsInstance.new(@aws, :worker, r[:key_name], r[:security_groups].map { |s| s[:group_id] }, @group_uuid, @private_key, @private_key_file_name, @proxy)
           @workers.last.load_instance_data(r)
         end
       end
     else
       logger.info 'Worker nodes are already defined'
     end
+
+    # set the private key from the hash
+    load_private_key server_data_hash[:server][:private_key_file_name]
+    load_worker_key server_data_hash[:server][:worker_private_key_file_name]
 
     # Really don't need to return anything because this sets the class instance variable
     @server
@@ -491,23 +587,23 @@ class OpenStudioAwsWrapper
     amis
   end
 
-  def to_os_worker_hash
-    worker_hash = []
-    @workers.each do |worker|
-      worker_hash.push(
+  # save off the instance configuration and instance information into a JSON file for later use
+  def to_os_hash
+    h = @server.to_os_hash
+
+    h[:server][:worker_private_key_file_name] = @worker_keys_filename
+    h[:workers] = @workers.map do |worker|
+      {
         id: worker.data.id,
         ip: "http://#{worker.data.ip}",
         dns: worker.data.dns,
         procs: worker.data.procs,
         private_key_file_name: worker.private_key_file_name,
         private_ip_address: worker.private_ip_address
-      )
+      }
     end
 
-    out = { workers: worker_hash }
-    logger.info out
-
-    out
+    h
   end
 
   # take the base version and increment the patch until
@@ -538,7 +634,7 @@ class OpenStudioAwsWrapper
       sv = ami[:tags_hash][:openstudio_server_version]
 
       if sv.nil? || sv == ''
-        puts 'found nil version, incrementing from 0.0.1'
+        logger.info 'found nil version, incrementing from 0.0.1'
         sv = get_next_version('0.0.1', list_of_svs)
       end
       list_of_svs << sv
@@ -590,10 +686,8 @@ class OpenStudioAwsWrapper
         end
       end
     end
-    # puts "Current AMIs: #{JSON.pretty_generate(amis)}"
 
     # merge in the existing AMIs the existing amis into the 'unknown category, but don't flag them as 'deprecate'
-    # puts "Existing AMIs: #{JSON.pretty_generate(existing)}"
     existing.keys.each do |ami_key|
       next if ami_key == 'default'.to_sym # ignore default
 
@@ -612,7 +706,6 @@ class OpenStudioAwsWrapper
       a[:openstudio_server_version] = next_version.to_s
     end
 
-    # puts "After merge: #{JSON.pretty_generate(amis)}"
     # flip these around for openstudio server section
     amis[:openstudio_server].keys.each do |key|
       next if key.to_s.to_version.satisfies('0.0.*')
@@ -638,8 +731,6 @@ class OpenStudioAwsWrapper
       amis[:openstudio][key][:default] = amis[:openstudio][key].keys[0]
     end
     amis[:openstudio] = Hash[amis[:openstudio].sort_by { |k, _| k.to_s.to_version }.reverse]
-
-    # puts "After sort: #{JSON.pretty_generate(amis)}"
 
     amis
   end